backend/requirements.txt
aio-pika==9.3.1
aiofiles==23.2.1
aiohttp==3.9.3
aiormq==6.7.7
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
aioamqp==0.15.0
backoff==2.2.1
bidict==0.22.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
DateTime==5.4
reportlab==4.2.2
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fpdf==1.7.2
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
googletrans==2.4.0
grpcio==1.60.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.15.0
httplib2==0.22.0
httpx==0.25.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.11.0
iniconfig==2.0.0
loguru==0.7.2
messagebird==2.1.0
mongomock==4.1.2
multidict==6.0.4
numpy==1.26.3
oauthlib==3.2.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
#opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-aio-pika==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-pymongo==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==23.2
pamqp==3.2.1
pandas==2.2.0
pdfkit==1.0.0
Jinja2==3.1.5
pillow==10.2.0
pillow-avif-plugin==1.4.6
pipreqs==0.4.13
pluggy==1.3.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
cryptography
moment
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
pypng==0.20220715.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-multipart==0.0.6
python-redsys==1.2.0
python-socketio==5.11.0
pytz==2023.3.post1
qrcode==7.4.2
redsys==0.2.7
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
sentinels==1.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
stripe==8.1.0
tomli==2.0.1
toolz==0.12.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
ua-parser==0.18.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
user-agents==2.2.0
utcnow==0.3.8
uvicorn==0.25.0
WooCommerce==3.0.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
